prevent downloading stream double
This could happen when changing the streamlist during a stream
recording:  A new goroutine would be spawned for the same stream, but
since the older goroutine was still recording it wouldn't be killed.
The new goroutine would start recording at the same time.

This fix makes the new goroutine wait until the previous one is finished
before beginning the check loop.